# openapi.yaml
openapi: 3.1.0
info:
  title: DataForGood API
  version: "1.0"
  description: |
    Route handlers de `app/api`. Las acciones del panel del SuperUsuario
    (`/sistema`, `/usuarios`, `/supervisar`) son server actions y no aparecen aquí.

    **Sesión de usuario:** se maneja con la cookie `session_token` (httpOnly). Ejecuta
    primero `POST /api/auth/login`; el navegador (o Insomnia) guarda la cookie
    y la manda sola en las siguientes peticiones.

    **SuperUsuario:** ejecuta `POST /api/auth/root/token` con la credencial de
    SuperUsuario del entorno (`ROOT_USER_ID` y su contraseña), copia `data.token`,
    pulsa **Authorize** y pégalo en `rootBearer`. Vale 2 horas. Con él responden
    `GET /api/usuarios` (todos los usuarios) y `GET /api/usuarios/{id}` (cualquier id).
    El resto de las rutas son de la cuenta de un usuario y piden `session_token`.

    **CSRF:** las mutaciones con un header `Origin` de otro sitio responden
    403 `Origen no permitido` (`proxy.ts`).

    Todas las respuestas JSON son `{ data }`, `{ message, data }` o `{ error }`.

    Este archivo se mantiene a mano: si cambias una ruta, actualízalo en el mismo commit.
# /api/docs/spec agrega al inicio el servidor desde el que se abre la página.
servers:
  - url: http://localhost:3000/dataforgood
    description: Docker local
  - url: http://localhost:3000
    description: next dev
  - url: https://multimodal-ai-lab.cicese.mx/dataforgood
    description: Producción
...
security:
  - sessionCookie: []
paths:
...
  /api/campanas/{id}/baneos:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Campañas]
      summary: Participantes baneados y dispositivos anónimos bloqueados (solo el creador)
      responses:
        "200":
          description: Baneados, del más reciente al más antiguo
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        usuarioId: { type: string }
                        nombre: { type: string }
                        motivo: { type: string }
                        baneadoEn: { type: string, format: date-time }
                  dispositivos:
                    type: array
                    description: Dispositivos de personas sin cuenta bloqueados en esta campaña
                    items:
                      type: object
                      properties:
                        bloqueoId: { type: string }
                        aporteId: { type: [string, "null"] }
                        motivo: { type: string }
                        bloqueadoEn: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Campañas]
      summary: Quitar el baneo de un participante o desbloquear un dispositivo anónimo (solo el creador)
      description: "Manda `usuarioId` para una cuenta o `bloqueoId` para un dispositivo. Bitácora: `campana.desbanear` o `campana.desbloquear_dispositivo`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuarioId: { type: integer }
                bloqueoId: { type: integer }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: La campaña no existe o esa persona no estaba baneada }
    post:
      tags: [Campañas]
      summary: Banear de la campaña a quien envió un aporte (solo el creador)
      description: |
        Si el aporte es de una cuenta, la banea de la campaña. Si es anónimo, bloquea en la
        campaña el dispositivo que lo envió (`campana.bloquear_dispositivo` en audit_log).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contributionId, reason]
              properties:
                contributionId: { type: string }
                reason: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/campanas/{id}/enlace:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Campañas]
      summary: Enlace público actual de la campaña
      description: |
        Cualquier usuario con sesión (el enlace es para difundir la campaña). Devuelve el último
        enlace no revocado, vigente o caducado, o `data: null` si nunca se generó. `visitas` y
        `aportesRecibidos` solo vienen si quien pide es el creador. El enlace se genera solo cuando
        la campaña queda `activa`.
      responses:
        "200":
          description: Enlace, estado de la campaña y si quien pide es el creador
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: [object, "null"]
                    properties:
                      url: { type: string }
                      token: { type: string }
                      creadoEn: { type: string, format: date-time }
                      expiraEn: { type: string, format: date-time }
                      vigente: { type: boolean }
                      visitas: { type: integer, description: Solo para el creador }
                      aportesRecibidos: { type: integer, description: Solo para el creador }
                  campana:
                    type: object
                    properties:
                      nombre: { type: string }
                      status: { type: string }
                      permiteAnonimos: { type: boolean, description: "Si /c/{token} acepta aportes sin cuenta" }
                  viewer:
                    type: object
                    properties:
                      esCreador: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Campañas]
      summary: Generar o regenerar el enlace público
      description: |
        Solo el creador y solo con la campaña `activa`. Crea un token que vale 24 horas
        para la ruta pública /c/{token}. Si ya había uno, queda revocado para siempre
        (conserva sus visitas y aportes). Se registra en audit_log como `campana.enlace_regenerar`.
      responses:
        "201":
          description: Enlace nuevo
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token: { type: string }
                      creadoEn: { type: string, format: date-time }
                      expiraEn: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: La campaña no está activa
    patch:
      tags: [Campañas]
      summary: Permitir o cortar los aportes sin cuenta
      description: Solo el creador. No toca los aportes anónimos ya recibidos. Bitácora `campana.anonimos_cambiar`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permiteAnonimos]
              properties:
                permiteAnonimos: { type: boolean }
      responses:
        "200":
          description: Valor guardado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object, properties: { permiteAnonimos: { type: boolean } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/campanas/{id}/qr:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Campañas]
      summary: Descargar el código QR del enlace público
      description: Cualquier usuario con sesión. QR de 512 × 512 px del enlace vigente, que apunta a /c/{token}.
      parameters:
        - name: formato
          in: query
          schema: { type: string, enum: [png, svg], default: png }
      responses:
        "200":
          description: Archivo del QR (como descarga)
          content:
            image/png: { schema: { type: string, format: binary } }
            image/svg+xml: { schema: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410":
          description: El enlace caducó; hay que regenerarlo

  # ---------------------------------------------------------------- Aportes
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
      description: Se obtiene con `POST /api/auth/login`. El navegador la manda sola.
    pendingVerificationCookie:
      type: apiKey
      in: cookie
      name: pending_verification_id
      description: La dejan el registro o un login sin verificar.
    rootBearer:
      type: http
      scheme: bearer
      description: Token del SuperUsuario. Se obtiene con `POST /api/auth/root/token` y vale 2 horas.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Message:
      description: Operación correcta
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
    BadRequest:
      description: Datos inválidos
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: No hay sesión (o la cuenta está suspendida o baneada)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Sin permiso para esta acción, o `Origen no permitido`
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: No existe
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...